requirements.txt
langchain-community 
langchain_openai
openai
fastapi
uvicorn
faiss-cpu
python-dotenv